flink-kafka-streaming/python-simulator/requirements.txt
kafka-python==2.0.2
faker==19.6.2
numpy==1.24.3